# OpenAPI.yaml
openapi: 3.0.0
info:
  title: Cart Management API
  version: 1.0.0
  description: API for managing user shopping carts and cart products.
servers:
  - url: 'http://localhost:8080'
paths:
  /carts/{userId}:
    post:
      summary: Create a new cart for a user
      operationId: createCart
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: New cart created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: Invalid input
        '404':
          description: User not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
  /carts/{id}:
    get:
      summary: Retrieve a user's cart and cart products
      operationId: getCart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: A cart object with a list of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cart not found for the specified user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a user's cart
      operationId: deleteCart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Cart deleted successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cart not found for the specified user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /carts/products:
    post:
      summary: Add a new product to the cart
      operationId: addProductToCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cart:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                productId:
                  type: integer
                  format: int64
                productName:
                  type: string
                productCategory:
                  type: string
                productDescription:
                  type: string
                quantity:
                  type: integer
                price:
                  type: number
                  format: double
      responses:
        '201':
          description: Product added successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update the quantity of an existing product in the cart
      operationId: updateQuantityProductInCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatedCartProductDto'
      responses:
        '200':
          description: Product quantity updated successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /carts/products/{id}:
    delete:
      summary: Remove a item from cart
      operationId: removeItemFromCart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartProducts'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Cart:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        updated_at:
          type: string
          format: date-time
        products:
          type: array
          items:
            $ref: '#/components/schemas/CartProducts'
            
    CartDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
          
    CartProducts:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cart_id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        productName:
          type: string
        productCategory:
          type: string
        productDescription:
          type: string
        quantity:
          type: integer
        price:
          type: number
          format: decimal
          
    ErrorResponse:
      type: object
      properties:
        code: 
          type: number
        message:
          type: string

    UpdatedCartProductDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        quantity:
          type: integer